Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pypdf==5.1.0
spacy==3.8.2
groq==0.11.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
